Image Services Created

+++ app/Http/Controllers/Admin/Blog/CategoryController.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\BaseController;
 use App\Http\Requests\CategoryStoreRequest;
 use App\Http\Requests\CategoryUpdateRequest;
 use App\Models\Category;
+use App\services\ImageServices;
 use App\Utlity;
 
 class CategoryController extends BaseController
@@ -12,6 +13,9 @@ class CategoryController extends BaseController
 
 
     public function index(){
+        if (is_null($this->user) || !$this->user->can('category.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any category !');
+        }
         return view('admin.pages.category.index', [
             'prefixname' => 'Admin',
             'title' => 'Category List',
@@ -22,6 +26,9 @@ class CategoryController extends BaseController
     }
 
     public function create(){
+        if (is_null($this->user) || !$this->user->can('category.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any category !');
+        }
         return view('admin.pages.category.create', [
             'prefixname' => 'Admin',
             'title' => 'Category Create',
@@ -30,16 +37,11 @@ class CategoryController extends BaseController
         ]);
     }
 
-    public function store(CategoryStoreRequest $request, Category $category){
-
-        //upload photo
-        if ($request->hasFile('img')){
-            $path['image'] = Utlity::file_upload($request,'img','Category_Photo');
+    public function store(CategoryStoreRequest $request, Category $category, ImageServices $imageServices){
+        if (is_null($this->user) || !$this->user->can('category.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any category !');
         }
-        else {
-            $path['image'] = null;
-        }
-
+        $path=$imageServices->imageStore('category');
         $data =array_merge($path, $request->only('nameBn','nameEn', 'description', 'status'));
         if ($category->create($data)) {
             return redirect()->route('category.index')->with('success', 'Data Added successfully Done');
@@ -48,6 +50,9 @@ class CategoryController extends BaseController
     }
 
     public function show(Category $category){
+        if (is_null($this->user) || !$this->user->can('category.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any category !');
+        }
         return view('admin.pages.category.edit', [
             'prefixname' => 'Admin',
             'title' => 'Category Show',
@@ -58,6 +63,9 @@ class CategoryController extends BaseController
     }
 
     public function edit(Category $category){
+        if (is_null($this->user) || !$this->user->can('category.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any category !');
+        }
         return view('admin.pages.category.edit', [
             'prefixname' => 'Admin',
             'title' => 'Category Edit',
@@ -66,18 +74,11 @@ class CategoryController extends BaseController
         ]);
     }
 
-    public function update(CategoryUpdateRequest $request,Category $category){
-
-
-        if($request->hasFile('img')){
-            if(file_exists($category->image)){
-                unlink($category->image);
-            }
-            $path['image'] = Utlity::file_upload($request,'img','Category_Photo');
-        }else{
-            $path['image']=null;
+    public function update(CategoryUpdateRequest $request,Category $category, ImageServices $imageServices){
+        if (is_null($this->user) || !$this->user->can('category.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any category !');
         }
-
+        $path=$imageServices->imageUpdate($category, 'category');
         $data =array_merge($path, $request->only('nameBn','nameEn', 'description', 'status'));
         if ($category->update($data)) {
             return redirect()->route('category.index', $category->id)->with('success', 'Data Updated successfully Done');
@@ -86,6 +87,9 @@ class CategoryController extends BaseController
     }
 
     public function destroy(Category $category){
+        if (is_null($this->user) || !$this->user->can('category.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any category !');
+        }
         if(file_exists($category->image)){
             unlink($category->image);
         }
